testing use cases: entire person use cases completed

+++ lib/domain/usecases/delete_person_by_id_usecase.dart
@@ -3,6 +3,7 @@ import 'package:all_4_u/core/usecases/usecase.dart';
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
 
+import '../../core/error/error_messages.dart';
 import '../repositories/person_repository_intf.dart';
 
 class DeletePersonByIdUsecase implements UseCase<void, DeletePersonParams> {
@@ -12,6 +13,10 @@ class DeletePersonByIdUsecase implements UseCase<void, DeletePersonParams> {
 
   @override
   Future<Either<Failure, void>> call(DeletePersonParams params) async {
+    if (params.props == null) {
+      // error - no params
+      return Future.value(Left(EmptyParams(errorMessage: PARAM_ERROR)));
+    }
     repository.deletePersonById(id: params.personId);
     return const Right(null);
   }
